+++ AppointmentService.Domain/CounselorScheduleSlot.cs
@@ -8,7 +8,7 @@ public partial class CounselorScheduleSlot
 
     public Guid ScheduleDayId { get; set; }
 
-    public short SlotId { get; set; }
+    public int SlotId { get; set; }
     
     public short Status { get; set; }
